backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "backboard-sdk>=1.5.3",
    "fastapi[standard]>=0.135.1",
    "google-genai>=1.66.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sqlmodel>=0.0.37",
    "uvicorn[standard]>=0.41.0",
]
